+++ packages/react-email/src/utils/constants.ts
@@ -1,6 +1,0 @@
-export const tabTransition = {
-  type: 'spring',
-  stiffness: 2000,
-  damping: 80,
-  mass: 1,
-};