+++ packages/react-email/src/utils/esbuild/renderring-utilities-exporter.ts
@@ -1,7 +1,7 @@
 import { promises as fs } from 'node:fs';
 import path from 'node:path';
 import type { Loader, PluginBuild, ResolveOptions } from 'esbuild';
-import { escapeStringForRegex } from './escape-string-for-regex';
+import { escapeStringForRegex } from './escape-string-for-regex.js';
 
 /**
  * Made to export the `render` function out of the user's email template