+++ packages/react-email/src/utils/testing/request-response-email.tsx
@@ -1,9 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-const req = new Request('https://react.email');
-const res = new Response('{}');
-
-const Email = () => {
-  return <div />;
-};
-
-export default Email;