+++ packages/react-email/src/utils/types/hot-reload-event.ts
@@ -1,3 +1,3 @@
-import type { EventName } from 'chokidar/handler';
+import type { EventName } from 'chokidar/handler.js';
 
 export type HotReloadEvent = EventName;